move file to api dir
