+++ src/utils/ipcComms.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, dialog, ipcMain, Tray, Notification } from 'electron';
 import { createWindow } from './createWindow';
 import { buildContextMenu } from './menuUtil';
 import { logErrorSentry } from './sentry';
-import { getFilesFromDir } from '../services/upload';
+import { getFilesFromDir } from '../api/upload';
 import chokidar from 'chokidar';
 import path from 'path';
